begin bluetooth refactor Move device discovery code to BleDiscovery class Split glukometr.py into separate files
