+++ common/assets/GlucoseApplication.qml
@@ -5,10 +5,18 @@ import io.thp.pyotherside 1.4
 Item {
     ListModel { id: measurementsModel }
     ListModel { id: thresholdsModel }
+    ListModel { id: rememberedDevicesModel }
+    ListModel { id: discoveredDevicesModel }
 
     property alias measurements: measurementsModel
     property alias thresholds: thresholdsModel
+    property alias rememberedDevices: rememberedDevicesModel
+    property alias discoveredDevices: discoveredDevicesModel
 
+    function addDevice(name, mac_address, remember) {
+        console.log(name, mac_address, remember)
+        python.call("glukometr.devices.add", [name, mac_address, remember, ]);
+    }
 
     function evaluateMeasurement(value, meal) {
         return python.call_sync("glukometr.thresholds.evaluate_measurement", [
@@ -75,14 +83,26 @@ Item {
             addImportPath(Qt.resolvedUrl('../python'));
             importModule('glukometr', function () {});
             getLastSequenceNumber(1);
+
+            setHandler("rememberedDevicesChanged", function (devices) {
+                setListModel(devices, rememberedDevices);
+            })
+            setHandler("discoveredDevicesChanged", function (devices) {
+                setListModel(devices, discoveredDevices);
+            })
+            loadListModel("glukometr.devices.get", rememberedDevices);
+        }
+
+        function setListModel(results, model) {
+            model.clear()
+            for (var i=0; i<results.length; i++) {
+                model.append(results[i]);
+            }
         }
 
         function loadListModel(pythonMethod, model) {
-            call(pythonMethod, [], function(result) {
-                model.clear()
-                for (var i=0; i<result.length; i++) {
-                    model.append(result[i]);
-                }
+            call(pythonMethod, [], function(results) {
+                setListModel(results, model);
             });
         }
