+++ sailfish/assets/home.qml
@@ -1,87 +1,122 @@
 import QtQuick 2.0
 import Sailfish.Silica 1.0
+import glukometr 1.0
+
 
 Page {
     id: screen
-    property string wiadomosc: glukometr.wiadomosc
 
-    Connections {
-        target: glukometr
-        onWiadomoscChanged:
-        {
-            busyIndicator.running = wiadomosc == "Skanowanie urządzeń"
-        }
-        onNazwaChanged: busyIndicator.running = false
-
+    BleDiscovery {
+        id: bleDiscovery
+        onNewDevice: pythonGlukometr.addDevice(name, macAddress, false)
+        Component.onCompleted: startDiscovery()
     }
 
-    Component.onCompleted: {
-        glukometr.urzadzenieSearch();
-    }
-
-    BusyIndicator {
-        id: busyIndicator
-        anchors.centerIn: parent
-        running: false
-        size: BusyIndicatorSize.Large
-
-        Label {
-            anchors {
-                horizontalCenter: busyIndicator.horizontalCenter
-                top: busyIndicator.bottom
-                topMargin: Theme.paddingLarge
-            }
-
-            text: glukometr.wiadomosc
-        }
-    }
-
-    SilicaListView {
-        id: theListView
-        width: parent.width
+    SilicaFlickable {
         anchors.fill: parent
-        model: glukometr.nazwa
+        contentHeight: column.height
 
-        header: PageHeader {
-            title: "Wybierz urządzenie"
-        }
+        Column {
+            id: column
+            width: screen.width
+            spacing: Theme.paddingLarge
 
-        delegate: Rectangle {
-            id: box
-            height:100
-            width: parent.width
-            color: "#2dc4c4"
-            border.color: "#000000"
-            border.width: 5
-            radius: 15
+            PageHeader { title: "Wybierz urządzenie" }
 
-            MouseArea {
-                anchors.fill: parent
-                onPressed: { box.color= "#209b9b"; box.height=110}
-                onClicked: {
-                    glukometr.connectToService(modelData.urzadzenieAdres);
-                    pageStack.push("monitor.qml");
+            SectionHeader { text: "Zapamiętane urządzenia" }
+            Repeater {
+                width: parent.width
+                model: pythonGlukometr.rememberedDevices
+
+                delegate: Rectangle {
+                    id: box
+                    height:100
+                    width: parent.width
+                    color: "#2dc4c4"
+                    radius: 15
+
+                    MouseArea {
+                        anchors.fill: parent
+                        onPressed: { box.color= "#209b9b"; box.height=110}
+                        onClicked: {
+                            glukometr.connectToService(mac_address);
+                            pageStack.push("monitor.qml");
+                        }
+                    }
+
+                    Label {
+                        font.pixelSize: Theme.fontSizeMedium
+                        text: name
+                        anchors.top: parent.top
+                        anchors.topMargin: 5
+                        anchors.horizontalCenter: parent.horizontalCenter
+                        color: "#000000"
+                    }
+
+                    Label {
+                        font.pixelSize: Theme.fontSizeTiny
+                        text: mac_address
+                        anchors.bottom: parent.bottom
+                        anchors.bottomMargin: 5
+                        anchors.horizontalCenter: parent.horizontalCenter
+                        color: "#000000"
+                    }
                 }
             }
 
-            Label {
-                id: urzadzenie
-                font.pixelSize: Theme.fontSizeMedium
-                text: modelData.urzadzenieNazwa
-                anchors.top: parent.top
-                anchors.topMargin: 5
-                anchors.horizontalCenter: parent.horizontalCenter
-                color: "#000000"
+            SectionHeader { text: "Wykryte urządzenia" }
+            ProgressBar {
+                 visible: bleDiscovery.running
+                 label: "Szukanie urządzeń"
+                 indeterminate: true
+             }
+
+            Button {
+                visible: !bleDiscovery.running
+                text: "Szukaj bro"
+                onClicked: bleDiscovery.startDiscovery()
             }
 
-            Label {
-                id: urzadzenieAdres
-                font.pixelSize: Theme.fontSizeTiny
-                text: modelData.urzadzenieAdres
-                anchors.bottom: parent.bottom
-                anchors.bottomMargin: 5
-                anchors.horizontalCenter: parent.horizontalCenter
-                color: "#000000"
+            Repeater {
+                model: pythonGlukometr.discoveredDevices
+
+                delegate: Rectangle {
+                    id: dBox
+                    height:100
+                    width: parent.width
+                    color: "#2dc4c4"
+                    border.color: "#000000"
+                    border.width: 5
+                    radius: 15
+
+                    MouseArea {
+                        anchors.fill: parent
+                        onPressed: { dBox.color= "#209b9b"; dBox.height=110}
+                        onClicked: {
+                            glukometr.connectToService(mac_address);
+                            pythonGlukometr.addDevice(name, mac_address, true)
+                            pageStack.push("monitor.qml");
+                        }
+                    }
+
+                    Label {
+                        font.pixelSize: Theme.fontSizeMedium
+                        text: name
+                        anchors.top: parent.top
+                        anchors.topMargin: 5
+                        anchors.horizontalCenter: parent.horizontalCenter
+                        color: "#000000"
+                    }
+
+                    Label {
+                        font.pixelSize: Theme.fontSizeTiny
+                        text: mac_address
+                        anchors.bottom: parent.bottom
+                        anchors.bottomMargin: 5
+                        anchors.horizontalCenter: parent.horizontalCenter
+                        color: "#000000"
+                    }
+                }
             }
         }
     }
