+++ sailfish/resources.qrc
@@ -3,6 +3,9 @@
         <file alias="__init__.py">../python/__init__.py</file>
         <file alias="glukometr.py">../python/glukometr.py</file>
         <file alias="storage.py">../python/storage.py</file>
+        <file alias="measurements.py">../python/measurements.py</file>
+        <file alias="thresholds.py">../python/thresholds.py</file>
+        <file alias="devices.py">../python/devices.py</file>
     </qresource>
     <qresource prefix="/">
         <file alias="assets/GlucoseApplication.qml">../common/assets/GlucoseApplication.qml</file>
